refactor: 리스트 렌더링 로직 추상화 및 ProductList 컴포넌트 재설계 (#1)
