+++ src/pages/SavingsCalculatorPage/SavingsCalculatorPage.tsx
@@ -4,9 +4,9 @@ import { useSavingsCalculator } from './hooks/useSavingsCalculator';
 
 import { PageHeader, Divider, Tabs } from 'components/common';
 import { default as SavingForm } from './components/SavingsForm';
+import { default as FilteredProducts } from './components/FilteredProducts';
 import { default as CalculationResult } from './components/CalculationResult';
 import { default as RecommendedProducts } from './components/RecommendedProducts';
-import { default as ProductList } from './components/ProductList';
 
 export default function SavingsCalculatorPage() {
   const {
@@ -32,20 +32,22 @@ export default function SavingsCalculatorPage() {
 
       <Divider borderHeight={16} spacingHeight={8} />
 
+      {/* 입력 결과 */}
       <section>
         <Tabs defaultValue="products">
           <Tabs.Panel label="적금 상품" value="products">
-            {/* 상품 목록 */}
-            <ProductList products={products} selectedProductId={selectedProductId} onSelect={handleProductSelect} />
+            <FilteredProducts
+              products={products}
+              selectedProductId={selectedProductId}
+              onSelect={handleProductSelect}
+            />
           </Tabs.Panel>
 
           <Tabs.Panel label="계산 결과" value="results">
-            {/* 계산 결과 */}
             <CalculationResult result={savingResult} />
 
             <Divider borderHeight={16} spacingHeight={8} />
 
-            {/* 추천 상품 */}
             <RecommendedProducts
               products={recommendedProducts}
               selectedProductId={selectedProductId}
